Logicas para tratar de produtos e pedidos

+++ src/main/java/com/application/pedidoapi/service/ProdutoService.java
@@ -1,11 +1,14 @@
 package com.application.pedidoapi.service;
 
+import com.application.pedidoapi.enums.Tipo;
 import com.application.pedidoapi.model.Pedido;
+import com.application.pedidoapi.model.PedidoItem;
 import com.application.pedidoapi.model.Produto;
 import com.application.pedidoapi.repository.ProdutoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -17,6 +20,9 @@ public class ProdutoService {
 
     @Autowired
     ProdutoRepository produtoRepository;
+
+    @Autowired
+    private PedidoItemService pedidoItemService;
 
     public Optional<Produto> save(Produto produto) {
         Optional<Produto> opProduto = Optional.ofNullable(produtoRepository.save(produto));
@@ -38,8 +44,12 @@ public class ProdutoService {
         return produtoRepository.findAll(pageable);
     }
 
-    public List<Produto> findByDescricao(String descricao) {
-        return produtoRepository.findByDescricao(descricao);
+    public List<Produto> findProdutoByDescricao(String descricao) {
+        return produtoRepository.findByDescricao(descricao, Tipo.PRODUTO);
+    }
+
+    public List<Produto> findServicoByDescricao(String descricao) {
+        return produtoRepository.findByDescricao(descricao, Tipo.SERVICO);
     }
 
     public Page<Produto> findAllByDescricaoPageable(String nomeDescricao, Pageable pageable) {
@@ -53,7 +63,12 @@ public class ProdutoService {
         }
         return Optional.empty();
     }
-    public boolean delete(Produto produto) {
+    public boolean delete(Produto produto, @Nullable List<PedidoItem> pedidosPendentes) {
+        List<PedidoItem> pedidoItens = pedidoItemService.findAllWithProduto(produto);
+        if(!pedidoItens.isEmpty()){
+            pedidosPendentes.addAll(pedidoItens);
+            return false;
+        }
         produtoRepository.delete(produto);
         Optional<Produto> deleteValidation = findById(produto.getId());
         if(!deleteValidation.isPresent()){
@@ -75,5 +90,10 @@ public class ProdutoService {
     }
 
 
-
+    public List<Produto> findAll(Tipo tipo) {
+        if (tipo.equals(Tipo.SERVICO)) {
+            return produtoRepository.findAllByTipo(Tipo.SERVICO);
+        }
+        return produtoRepository.findAllByTipo(Tipo.PRODUTO);
+    }
 }
